refactor: Rename design reducer to theme
